+++ packages/dds/tree/src/shared-tree-core/editManagerSummaryFormatV1ToV2.ts
